Move SqlFunctionCallBase class into its own file
